Move ping endpoint to new module.
